Move enumeration structure into mod
